Fix window is not defined when starting the desktop app
After #6736 the theme was being called too early because of a chain
of calls from CalendarModel -> CalendarUtils -> CalendarGuiUtils -> ...
-> theme -> DeviceConfig throwing an error inside deviceConfig.

This commit moves the function canFullyEditEvent to CalendarViewModel
making sure that all the GUI stuff have been already loaded.

fix #6795
